requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.3
django-extensions==4.1
et_xmlfile==2.0.0
executing==2.2.0
html5lib==1.1
idna==3.10
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
lxml==6.0.0
matplotlib-inline==0.1.7
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pypdf==5.8.0
python-bidi==0.6.6
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
razorpay==1.4.2
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
traitlets==5.14.3
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
